Remove nested if and used a conjuct instead

+++ tally_ho/apps/tally/models/ballot.py
@@ -120,10 +120,9 @@ def auto_delete_document(sender, instance, **kwargs):
         return False
 
     new_document = instance.document
-    if old_document:
-        if old_document != new_document:
-            if os.path.isfile(old_document.path):
-                os.remove(old_document.path)
+    if old_document and old_document != new_document:
+        if os.path.isfile(old_document.path):
+            os.remove(old_document.path)
     return False
